Plugin hidden on start page
When we are on a start page with no other file, the plugin is hidden.
It’d be usefull to find a solution. Why not the plugin should remain
always visible? I’d be more simple. The images "-off" seem to be
sufficients to indicate no other page. In addition, the layout is thus
more homogeneous.